Move experimental commented out code to worlds.js too
